+++ Source/LuaBridge/detail/LuaException.h
@@ -7,24 +7,19 @@
 #pragma once
 
 #include "Config.h"
+
 #include "LuaHelpers.h"
 
+#include <cassert>
 #include <string>
 #include <sstream>
-
-#if LUABRIDGE_HAS_EXCEPTIONS
 #include <exception>
-#endif
 
 namespace luabridge {
 
+//================================================================================================
 class LuaException : public std::exception
 {
-private:
-    lua_State* m_L = nullptr;
-    int m_code = LUA_ERRERR;
-    std::string m_what;
-
 public:
     //=============================================================================================
     /**
@@ -32,7 +27,7 @@ public:
      *
      * Assumes the error string is on top of the stack, but provides a generic error message otherwise.
      */
-    LuaException(lua_State* L, int code)
+    LuaException(lua_State* L, std::error_code code)
         : m_L(L)
         , m_code(code)
     {
@@ -59,31 +54,15 @@ public:
      * This centralizes all the exceptions thrown, so that we can set breakpoints before the stack is unwound, or otherwise customize the behavior.
      */
     template <class Exception>
-    static void raise(lua_State* L, const Exception& e)
+    static void raise(const Exception& e)
     {
-#if LUABRIDGE_HAS_EXCEPTIONS
-        (void) L;
+        assert(areExceptionsEnabled());
 
+#if LUABRIDGE_HAS_EXCEPTIONS
         throw e;
 #else
-        luaL_error(L, "%s", e);
+        std::abort();
 #endif
-    }
-    
-    //=============================================================================================
-    /**
-     * @brief Wrapper for lua_pcall that throws if exceptions are enabled.
-     */
-    static int pcall(lua_State* L, int nargs = 0, int nresults = 0, int msgh = 0)
-    {
-        int code = lua_pcall(L, nargs, nresults, msgh);
-
-#if LUABRIDGE_HAS_EXCEPTIONS
-        if (code != LUABRIDGE_LUA_OK)
-            raise(L, LuaException(L, code));
-#endif
-        
-        return code;
     }
 
     //=============================================================================================
@@ -92,12 +71,24 @@ public:
      *
      * Subsequent Lua errors are translated to C++ exceptions, or logging and abort if exceptions are disabled.
      */
+    static bool areExceptionsEnabled() noexcept
+    {
+        return exceptionsEnabled();
+    }
+
+    /**
+     * @brief Initializes error handling.
+     *
+     * Subsequent Lua errors are translated to C++ exceptions, or logging and abort if exceptions are disabled.
+     */
     static void enableExceptions(lua_State* L) noexcept
     {
+        exceptionsEnabled() = true;
+        
         lua_atpanic(L, panicHandlerCallback);
     }
 
-protected:
+private:
     void whatFromStack()
     {
         std::stringstream ss;
@@ -105,34 +96,49 @@ protected:
         const char* errorText = nullptr;
         
         if (lua_gettop(m_L) > 0)
+        {
             errorText = lua_tostring(m_L, -1);
+            lua_pop(m_L, 1);
+        }
 
-        ss << (errorText ? errorText : "Unknown error") << " (code=" << m_code << ")";
+        ss << (errorText ? errorText : "Unknown error") << " (code=" << m_code.message() << ")";
 
         m_what = std::move(ss).str();
     }
 
-private:
     static int panicHandlerCallback(lua_State* L)
     {
 #if LUABRIDGE_HAS_EXCEPTIONS
-        throw LuaException(L, -1);
+        throw LuaException(L, makeErrorCode(ErrorCode::LuaFunctionCallFailed));
 #else
-        writestringerror("Unprotected error in call (%s)", lua_tostring(L, -1));
-        return 0;
+        std::abort();
 #endif
     }
+
+    static bool& exceptionsEnabled()
+    {
+        static bool areExceptionsEnabled = false;
+        return areExceptionsEnabled;
+    }
+
+    lua_State* m_L = nullptr;
+    std::error_code m_code;
+    std::string m_what;
 };
 
 //=================================================================================================
 /**
- * @brief Initializes error handling.
+ * @brief Initializes error handling using C++ exceptions.
  *
- * Subsequent Lua errors are translated to C++ exceptions, or logging and abort if exceptions are disabled.
+ * Subsequent Lua errors are translated to C++ exceptions. It aborts the application if called when no exceptions.
  */
 inline void enableExceptions(lua_State* L) noexcept
 {
+#if LUABRIDGE_HAS_EXCEPTIONS
     LuaException::enableExceptions(L);
+#else
+    assert(false); // Never call this function when exceptions are not enabled.
+#endif
 }
 
 } // namespace luabridge