+++ Tests/Source/LuaRefTests.cpp
@@ -54,8 +54,9 @@ TEST_F(LuaRefTests, ValueAccess)
     EXPECT_TRUE(result().isFunction());
     auto fnResult = result()(41); // Replaces result variable
     EXPECT_TRUE(fnResult);
-    EXPECT_TRUE(fnResult->isNumber());
-    ASSERT_EQ(41, fnResult->cast<int>());
+    EXPECT_TRUE(fnResult.getNumValues());
+    EXPECT_TRUE(fnResult.getValue(0).isNumber());
+    ASSERT_EQ(41, fnResult.getValue(0).cast<int>());
     EXPECT_TRUE(result().isNumber());
     ASSERT_EQ(42, result<int>());
 }
@@ -104,8 +105,9 @@ TEST_F(LuaRefTests, DictionaryRead)
     EXPECT_TRUE(result()["fn"].isFunction());
     auto fnResult = result()["fn"](41); // Replaces result variable
     EXPECT_TRUE(fnResult);
-    EXPECT_TRUE(fnResult->isNumber());
-    ASSERT_EQ(41, fnResult->cast<int>());
+    EXPECT_TRUE(fnResult.getNumValues());
+    EXPECT_TRUE(fnResult.getValue(0).isNumber());
+    ASSERT_EQ(41, fnResult.getValue(0).cast<int>());
     EXPECT_TRUE(result().isNumber());
     ASSERT_EQ(42, result<int>());
 }